+++ sickbeard/server/core.py
@@ -10,8 +10,8 @@ from tornado.web import Application, StaticFileHandler, RedirectHandler
 import sickbeard
 from sickbeard import logger
 from sickbeard.helpers import create_https_certificates, generateApiKey
-from sickbeard.server.web import LoginHandler, LogoutHandler, KeyHandler, CalendarHandler
-from sickbeard.server.api import ApiHandler
+from sickbeard.server.api.core import ApiHandler
+from sickbeard.server.web.core import LoginHandler, LogoutHandler, KeyHandler, CalendarHandler
 from sickrage.helper.encoding import ek
